Fix stopwatch pausing and stopping issues
Stopwatch was getting paused when tab was switched because the
browsers pause setInterval below 1000ms from running when not in
focus. This has now been fixed using different approach to calculate
elapsedTime for stopwatch.
Stopwatch also used to lose its state when the tool was switched to
clock because of component unmount. Now component is not unmounted
when tool is switched but it is hidden or shown depending on the
mode.